chore - split the direct-rustc artifact, inspection and diagnostic shapes (#1299)
`src/oven/rustc.rs` carries ~6,700 production lines. This is the first of the
per-file splits #1299 describes, and it moves only: no logic is edited, and every
public path stays where callers expect it through re-exports, so nothing outside
this module changes.

Three seams the file already had, taken as they were:

- `artifact` -- what a direct-rustc invocation compiles and links: externs,
  supporting and auxiliary artifacts, the manifest recording them, and the
  registry leaves a sealed plan resolves against.
- `inspection` -- the sealed project-inspection authority payload: sources,
  constituents, root and test dependencies, generated output directory.
- `diagnostics` -- rustc's JSON spans and messages, and the report that renders
  them beside the invocation that produced them.

Two adjustments were forced by the move and are the only non-move edits. Imports
each block needs are now declared where it lives rather than inherited from one
shared list. Two fields of the loaded inspection authority became `pub(super)`,
which is the reach they already had when the struct shared a module with its
only reader; nothing gained visibility beyond that module.

Evidence it is a move: the exported item names are identical before and after --
85 either way, none lost, none gained -- and `cargo test --lib` reports the same
3234 passed / 28 failed on this branch as on the unmodified branch, those 28
being the known `CARGO_BIN_EXE_incan` gap that plain `--lib` runs always hit.

`legacy_cargo.rs` and `cli/commands/oven.rs` remain, one per pull request as the
issue prescribes.
