+++ src/oven/rustc.rs
@@ -4,10 +4,20 @@
 //! An explicit publisher-side `legacy_cargo` step may create the declared inputs, but this consumer path invokes only
 //! the selected Rust compiler and refuses hidden Cargo state.
 
+mod artifact;
+mod diagnostics;
+mod inspection;
+
+// Split along seams this file already had: the artifact/plan shapes, the project-inspection authority payloads,
+// and the rustc diagnostic report. Every path stays where callers expect it -- re-exported here rather than
+// re-homed -- so this is a move, not an interface change.
+pub use artifact::*;
+pub use diagnostics::*;
+pub(crate) use inspection::*;
+
 use std::collections::{BTreeMap, BTreeSet};
 use std::env;
 use std::ffi::OsString;
-use std::fmt;
 use std::fs;
 use std::io::{self, Read};
 use std::path::{Component, Path, PathBuf};
@@ -39,368 +49,6 @@ pub const OVEN_RUSTC_REGISTRY_LOCK_RELATIVE_PATH: &str = "registry-sources/Cargo
 pub const OVEN_COMPILER_RUNTIME_CRATE_PREFIX: &str = "incan_";
 /// Schema version for caller-owned native-output reuse evidence.
 const OVEN_DIRECT_RUSTC_OUTPUT_RECEIPT_SCHEMA_VERSION: u32 = 2;
-
-/// One exact dependency artifact permitted in an Oven direct-rustc invocation.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcArtifactExtern {
-    /// Crate name passed to `rustc --extern`.
-    pub crate_name: String,
-    /// Safe path relative to the immutable artifact root.
-    pub relative_path: String,
-    /// SHA-256 digest of the exact selected artifact bytes.
-    pub digest: String,
-}
-
-/// One non-root artifact reachable through a declared dependency or native search directory.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcSupportingArtifact {
-    /// Safe path relative to the immutable artifact root.
-    pub relative_path: String,
-    /// SHA-256 digest of the exact selected artifact bytes.
-    pub digest: String,
-}
-
-/// One compiler-owned cross-target Rust closure reserved for vocabulary desugarer packaging.
-///
-/// Normal Oven commands can compile caller-owned vocabulary companions to this target, but they cannot add targets
-/// or select arbitrary dependency artifacts. The named publisher has already resolved, copied, and digested every
-/// file named here.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcAuxiliaryTarget {
-    /// Rust target triple for this isolated closure.
-    pub target: String,
-    /// Safe artifact-root-relative directories passed as `-L dependency` for this target only.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub dependency_search_paths: Vec<String>,
-    /// Exact target-specific dependency roots passed through `--extern`.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub externs: Vec<OvenRustcArtifactExtern>,
-}
-
-/// Immutable direct-rustc dependency plan for one Oven compatibility domain.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcArtifactManifest {
-    /// Manifest wire-schema version.
-    pub schema_version: u32,
-    /// Target/toolchain/profile/features that own the selected artifacts.
-    pub intent: OvenBuildIntent,
-    /// Safe artifact-root-relative directories passed as `-L dependency`.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub dependency_search_paths: Vec<String>,
-    /// Safe artifact-root-relative directories passed as `-L native`.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub native_search_paths: Vec<String>,
-    /// Exact root dependency artifacts passed through `--extern`.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub externs: Vec<OvenRustcArtifactExtern>,
-    /// Exact root extern names authorized for each receipt source-evidence key.
-    ///
-    /// A package library test and its CLI can share one immutable dependency closure while requiring different direct
-    /// `--extern` roots. The selected list prevents one target's test-only direct dependency from overriding another
-    /// target's metadata-selected transitive dependency instance.
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub entrypoint_externs: BTreeMap<String, Vec<String>>,
-    /// Exact registry package artifacts whose metadata closure was emitted with this immutable plan.
-    ///
-    /// The Loaf repeats this catalog for human inspection, while this copy travels with every bounded
-    /// store entry so a selected plan resolves caller `rust::` imports only from its own compatibility domain.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub registry_leaves: Vec<OvenRustcRegistryLeaf>,
-    /// Complete locked registry source closure authorized for build-system-neutral Rust inspection.
-    ///
-    /// This is distinct from `registry_leaves`: a transitive proc macro may be required by the source graph without
-    /// exposing an `.rlib` that a caller can select as a direct Rust dependency.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub registry_sources: Vec<OvenRustcRegistrySourcePackage>,
-    /// Deterministic compile-time environment explicitly required by the source closure.
-    ///
-    /// Ambient `CARGO_*` values are still removed before every consumer invocation. The only permitted replacements
-    /// are package metadata captured by the publisher and `@oven-source-root`, which resolves to the caller-owned
-    /// root derived from the receipt-authorized source path so compatible clean worktrees do not inherit another
-    /// checkout's location.
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub compile_environment: BTreeMap<String, String>,
-    /// Compiler-owned target closures used only while packaging vocabulary Wasm desugarers.
-    ///
-    /// These are deliberately separate from the normal host plan: placing cross-target artifacts on every caller
-    /// compilation's dependency search path would make Rustc's artifact selection target-ambiguous.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub vocab_auxiliary_targets: Vec<OvenRustcAuxiliaryTarget>,
-    /// Every other regular artifact reachable through declared search paths.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub supporting_artifacts: Vec<OvenRustcSupportingArtifact>,
-}
-
-/// Materialized, content-verified inputs ready for a direct `rustc` command.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct OvenRustcArtifactPlan {
-    /// Verified directories passed as `-L dependency`.
-    pub dependency_search_paths: Vec<PathBuf>,
-    /// Verified directories passed as `-L native`.
-    pub native_search_paths: Vec<PathBuf>,
-    /// Verified root dependency artifacts passed as `--extern`.
-    pub externs: Vec<(String, PathBuf)>,
-    /// Verified compiler-owned environment replacements applied only while compiling the caller-owned source.
-    pub compile_environment: BTreeMap<String, String>,
-    /// Digests of caller-owned direct-Rustc libraries linked in addition to immutable Oven artifacts.
-    ///
-    /// A compiler-suite receipt covers its workspace source closure, but output reuse also records these exact
-    /// library bytes so a changed intermediate cannot be mistaken for the previously linked root.
-    pub(crate) caller_owned_library_digests: BTreeMap<String, String>,
-}
-
-/// A verified partition of one complete direct-Rustc closure across a selected base Loaf and a project extension.
-///
-/// The partition is derived from exact relative paths and digests, never crate names.  This makes an extension a
-/// genuine delta: a same-named artifact whose bytes differ is not silently borrowed from a base compiled with a
-/// different feature graph.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct OvenRustcArtifactPartition {
-    /// Artifacts supplied by the selected immutable base Loaf.
-    pub(crate) base_paths: BTreeSet<String>,
-    /// Artifacts that must be retained in the receipt-bound extension Loaf.
-    pub(crate) extension_paths: BTreeSet<String>,
-}
-
-/// Materialized compiler-owned cross-target closure available only to vocabulary extraction.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct OvenRustcAuxiliaryTargetPlan {
-    pub(crate) dependency_search_paths: Vec<PathBuf>,
-    pub(crate) externs: Vec<(String, PathBuf)>,
-}
-
-/// One caller-owned library or procedural-macro artifact made by an earlier Oven direct-Rustc materialization step.
-///
-/// This is deliberately an internal compiler-suite bridge, not an artifact-store entry. The library remains below
-/// the caller's output root and carries its already-verified digest into the later output's reuse identity.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct OvenCallerOwnedRustcLibrary {
-    /// Rust crate name used by a consuming direct-Rustc root's `--extern` flag.
-    pub crate_name: String,
-    /// Regular caller-owned `.rlib` or dynamic procedural-macro output from the earlier direct-Rustc step.
-    pub output: PathBuf,
-    /// Digest established when Oven baked or revalidated this caller-owned output in the current execution.
-    ///
-    /// The output is created below this invocation's caller-owned directory, then its bytes are read once before it
-    /// becomes a dependency. Later consumers reuse that verified identity rather than re-reading the same large
-    /// workspace library for every downstream target.
-    pub digest: String,
-    /// Whether this output is a direct dependency of the current compilation root.
-    ///
-    /// Transitive caller-owned outputs still need a verified `-L dependency` search path and reuse evidence, but
-    /// must not be presented as arbitrary public `--extern` names to the final consumer. That preserves the
-    /// provider's receipt-authorized visibility graph instead of flattening it into the consuming package.
-    pub expose_extern: bool,
-}
-
-/// One publisher-sealed registry package artifact that a Loaf may expose to a direct-Rustc consumer.
-///
-/// The catalog records an exact package version and the artifact Cargo emitted while the named publisher prepared the
-/// Loaf. It is not a general resolver: a consumer may select only one compatible record already copied into
-/// the immutable Loaf closure.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcRegistryLeaf {
-    /// Registry package name rather than a caller-local dependency alias.
-    pub package: String,
-    /// Exact publisher-resolved package version.
-    pub version: String,
-    /// Crate identifier encoded in the sealed Rust artifact metadata.
-    pub crate_name: String,
-    /// Publisher-resolved Cargo features compiled into this exact immutable artifact.
-    ///
-    /// A consumer may request only a subset. This represents the already unified Loaf closure; it does not
-    /// run a feature resolver or permit a consumer to add a feature absent from the sealed leaf.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub features: Vec<String>,
-    /// Exact registry source closure retained for build-system-neutral Rust inspection.
-    pub source: OvenRustcRegistrySource,
-    /// Digest-verified compiler artifact retained below the Loaf root.
-    pub artifact: OvenRustcArtifactExtern,
-}
-
-/// Publisher-sealed registry source corresponding to one compiled registry leaf.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcRegistrySource {
-    /// Cargo registry identity recorded by the exact publisher lock.
-    pub registry: String,
-    /// Registry archive checksum recorded by the exact publisher lock.
-    pub checksum: String,
-    /// Safe directory path relative to the immutable Loaf root.
-    pub relative_root: String,
-    /// Content digest of every regular source file and its portable relative path.
-    pub digest: String,
-}
-
-/// One exact locked registry package in the sealed Rust-inspection source closure.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcRegistrySourcePackage {
-    /// Cargo package name recorded by the explicit publisher.
-    pub package: String,
-    /// Exact package version recorded by the explicit publisher lock.
-    pub version: String,
-    /// Unified feature set selected by the publisher's resolved graph.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub features: Vec<String>,
-    /// Immutable source identity and artifact-root-relative location.
-    pub source: OvenRustcRegistrySource,
-}
-
-/// Wire schema for one project-level Rust inspection authority.
-pub(crate) const OVEN_PROJECT_INSPECTION_AUTHORITY_SCHEMA_VERSION: u32 = 1;
-
-/// Exact immutable source owner named by a project inspection authority.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "owner", rename_all = "snake_case")]
-pub(crate) enum OvenProjectInspectionSourceOwner {
-    /// The small project authority entry materializes this source fragment itself.
-    Authority,
-    /// One exact constituent supplies the source tree at the catalog's relative root.
-    Constituent { index: usize },
-}
-
-/// One exact immutable constituent of a project inspection authority.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub(crate) enum OvenProjectInspectionConstituent {
-    /// A compiler-shipped release Loaf retained by its immutable toolchain generation.
-    ReleaseLoaf {
-        loaf_identity: String,
-        build_unit_identity: String,
-        receipt: OvenReceipt,
-    },
-    /// A receipt-bound entry in the bounded project store.
-    Stored {
-        identity: String,
-        artifact_kind: OvenArtifactKind,
-        receipt: OvenReceipt,
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        base_loaf_identity: Option<String>,
-    },
-}
-
-/// One canonical registry source and the exact immutable root that owns its bytes.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct OvenProjectInspectionSource {
-    pub package: OvenRustcRegistrySourcePackage,
-    pub owner: OvenProjectInspectionSourceOwner,
-}
-
-/// One exact normal or dev registry root selected for project Rust inspection.
-///
-/// The locked package identity proves which source tree Cargo selected at the explicit bake boundary. The requested
-/// feature contract remains separate because two source-identical root edges can expose different Rust APIs.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct OvenProjectInspectionRootDependency {
-    pub alias: String,
-    pub package: String,
-    pub version: String,
-    pub registry: String,
-    pub checksum: String,
-    pub requested_features: Vec<String>,
-    pub default_features: bool,
-}
-
-/// Exact project-owned dependency envelope used only by generated native tests.
-///
-/// The envelope promotes the canonical normal and dev dependency surface into one checked debug executable closure at
-/// the explicit project-bake boundary. Its dependency digest is deliberately independent of authored test bytes, so
-/// unchanged dependency declarations reuse the same Loaf while every generated harness remains caller-owned.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct OvenProjectInspectionTestDependencyEnvelope {
-    pub constituent_index: usize,
-    pub dependency_surface_digest: String,
-    pub dependency_roots: BTreeMap<String, OvenProjectInspectionTestDependencyRoot>,
-}
-
-/// Exact per-root evidence admitted by the generated native-test dependency envelope.
-///
-/// Registry roots retain Cargo's locked package/source identity as well as the declared edge digest. Path and Git
-/// roots carry their complete portable declaration/source digest; for paths that digest includes the source tree,
-/// never its machine-local spelling.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "source", rename_all = "snake_case")]
-pub(crate) enum OvenProjectInspectionTestDependencyRoot {
-    Registry {
-        dependency_digest: String,
-        locked: OvenProjectInspectionRootDependency,
-    },
-    Path {
-        dependency_digest: String,
-    },
-    Git {
-        dependency_digest: String,
-    },
-}
-
-/// Singular source authority published once by an explicit project bake.
-///
-/// The authority owns one canonical publisher lock and exact normal/dev root-edge records. Its source catalog is a
-/// composition of named immutable constituents plus only those bounded source fragments absent from every
-/// constituent; normal commands never union independent locks or search the store by dependency compatibility.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct OvenProjectInspectionAuthorityPayload {
-    pub schema_version: u32,
-    pub project_identity: String,
-    pub source_authority_digest: String,
-    pub compiler_version: String,
-    pub registry_lock_digest: String,
-    #[serde(default)]
-    pub registry_source_dependencies: Vec<OvenProjectInspectionRootDependency>,
-    #[serde(default)]
-    pub dev_registry_source_dependencies: Vec<OvenProjectInspectionRootDependency>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub test_dependency_envelope: Option<OvenProjectInspectionTestDependencyEnvelope>,
-    #[serde(default)]
-    pub constituents: Vec<OvenProjectInspectionConstituent>,
-    #[serde(default)]
-    pub registry_sources: Vec<OvenProjectInspectionSource>,
-    /// Build-script output directories the explicit bake sealed below this authority's artifact root.
-    ///
-    /// Generated Rust such as prost's `include!`d modules exists as files only where the bake's Cargo bootstrap
-    /// wrote its `OUT_DIR`s. A normal command never runs Cargo, so the authority carries those files itself and a
-    /// direct inspection workspace reads them the way it would read a Cargo `OUT_DIR`.
-    #[serde(default)]
-    pub generated_out_dirs: Vec<OvenProjectInspectionGeneratedOutDir>,
-}
-
-/// One sealed build-script output directory, laid out as `generated-out-dirs/build/<crate>-<hash>/out` so the
-/// generated-code route recognizes it like a Cargo target directory.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct OvenProjectInspectionGeneratedOutDir {
-    /// Cargo package whose build script produced the directory.
-    pub crate_name: String,
-    /// Directory below the authority's artifact root holding the sealed `*.rs` output.
-    pub relative_root: String,
-    /// Exact package version whose build script wrote the directory, when the bake knew it. A closure can hold
-    /// several build units of one package, and a consumer reads only the unit built from the version it inspects.
-    #[serde(default)]
-    pub version: Option<String>,
-}
-
-/// Exact authority entry named by a source-current completed project output.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(crate) struct OvenProjectInspectionAuthorityRef {
-    pub identity: String,
-    pub receipt_identity: String,
-    pub build_unit_identity: String,
-}
-
-/// Source-current singular project authority with every bounded-store constituent leased in one batch.
-pub(crate) struct OvenLoadedProjectInspectionAuthority {
-    pub(crate) identity: String,
-    pub(crate) artifact_root: PathBuf,
-    pub(crate) payload: OvenProjectInspectionAuthorityPayload,
-    pub(crate) stored_constituents: Vec<OvenStoreExecutionPayload>,
-    _authority_lease: OvenStoreLease,
-    lineage_leases: Vec<OvenStoreLease>,
-}
-
-impl OvenLoadedProjectInspectionAuthority {
-    /// Retain completed-output leases for the complete inspection command.
-    pub(crate) fn retain_lineage_leases(&mut self, leases: Vec<OvenStoreLease>) {
-        self.lineage_leases = leases;
-    }
-}
 
 /// One registry leaf and the immutable Loaf root that seals its relative artifact path.
 #[derive(Debug, Clone)]
@@ -1919,110 +1567,6 @@ pub(crate) struct OvenTrustedRustdocTestRequest<'a> {
 pub(crate) struct OvenRustdocTestReport {
     /// Combined Rustdoc/doctest output for caller failure reporting.
     pub output: String,
-}
-
-/// Structured source span emitted by rustc's JSON diagnostic stream.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcDiagnosticSpan {
-    /// Rust source filename from rustc.
-    pub file_name: String,
-    /// One-based start line.
-    pub line_start: u32,
-    /// One-based start column.
-    pub column_start: u32,
-    /// One-based end line.
-    pub line_end: u32,
-    /// One-based end column.
-    pub column_end: u32,
-    /// Whether rustc identified this as the primary span.
-    pub is_primary: bool,
-}
-
-/// One structured rustc diagnostic preserved without terminal-only parsing.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcDiagnostic {
-    /// rustc severity level.
-    pub level: String,
-    /// Human-readable diagnostic message.
-    pub message: String,
-    /// Optional rustc error code.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub code: Option<String>,
-    /// Structured source spans.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub spans: Vec<OvenRustcDiagnosticSpan>,
-    /// Optional rustc-rendered display form.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub rendered: Option<String>,
-}
-
-/// Rustc diagnostic transcript for a failed direct Oven consumer compilation.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct OvenRustcDiagnosticReport {
-    /// JSON diagnostics decoded from rustc output.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub diagnostics: Vec<OvenRustcDiagnostic>,
-    /// Non-JSON rustc output retained verbatim for diagnostics that lack a JSON record.
-    #[serde(default, skip_serializing_if = "String::is_empty")]
-    pub unstructured_output: String,
-    /// Bounded direct-Rustc command evidence for a failed Oven compilation.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub invocation: Option<String>,
-}
-
-impl fmt::Display for OvenRustcDiagnosticReport {
-    /// Render a bounded, actionable terminal summary while the complete structured report remains available to callers.
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        const MAX_DIAGNOSTICS: usize = 12;
-        const MAX_UNSTRUCTURED_CHARS: usize = 4_000;
-
-        if self.diagnostics.is_empty() && self.unstructured_output.trim().is_empty() && self.invocation.is_none() {
-            return formatter.write_str("rustc exited unsuccessfully without emitting diagnostics");
-        }
-        for (index, diagnostic) in self.diagnostics.iter().take(MAX_DIAGNOSTICS).enumerate() {
-            if index > 0 {
-                formatter.write_str("\n")?;
-            }
-            write!(formatter, "{}", diagnostic.level)?;
-            if let Some(code) = &diagnostic.code {
-                write!(formatter, "[{code}]")?;
-            }
-            write!(formatter, ": {}", diagnostic.message)?;
-            if let Some(span) = diagnostic.spans.iter().find(|span| span.is_primary) {
-                write!(
-                    formatter,
-                    " at {}:{}:{}",
-                    span.file_name, span.line_start, span.column_start
-                )?;
-            }
-        }
-        if self.diagnostics.len() > MAX_DIAGNOSTICS {
-            write!(
-                formatter,
-                "\n… {} additional rustc diagnostic(s) omitted from terminal summary",
-                self.diagnostics.len() - MAX_DIAGNOSTICS
-            )?;
-        }
-        let unstructured = self.unstructured_output.trim();
-        if !unstructured.is_empty() {
-            if !self.diagnostics.is_empty() {
-                formatter.write_str("\n")?;
-            }
-            for character in unstructured.chars().take(MAX_UNSTRUCTURED_CHARS) {
-                write!(formatter, "{character}")?;
-            }
-            if unstructured.chars().count() > MAX_UNSTRUCTURED_CHARS {
-                formatter.write_str("\n… rustc unstructured output truncated")?;
-            }
-        }
-        if let Some(invocation) = &self.invocation {
-            if !self.diagnostics.is_empty() || !unstructured.is_empty() {
-                formatter.write_str("\n")?;
-            }
-            write!(formatter, "direct rustc invocation: {invocation}")?;
-        }
-        Ok(())
-    }
 }
 
 /// Successful direct-rustc consumer compilation evidence.
